chore: moved crawler check to its own pkg
